add stored fields and filtering of results
This allows to easily return the value of some fields in the search
results, and allows filtering of results of search or auto suggestions.

In particular, filtering auto suggestions on the basis of, say, the
document category, was not possible before this. Filtering of search
results was possible, but had to be done outside of MiniSearch.